+++ test/server/fixtures/data-params-project/client/parent.js
@@ -1,0 +1,7 @@
+import { Template } from "meteor/templating";
+
+Template.parent.helpers({
+    computedSubtitle() {
+        return "sub";
+    },
+});